requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.12.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
CacheControl==0.14.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dateparser==1.1.8
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.3.1
firebase_admin==7.3.0
frozenlist==1.3.3
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.4
msgpack==1.1.2
multidict==6.0.4
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.21
pycryptodome==3.17
PyJWT==2.12.1
pymongo==4.3.3
python-binance==1.0.17
python-dateutil==2.8.2
python-telegram==0.13.0
python-telegram-bot==13.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
six==1.16.0
telegram==0.0.1
tornado==6.3.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==1.26.20
websockets==11.0.3
yarl==1.9.2
